chore: improvement sweep - bugs, refactoring, code quality
Bugs fixed:
- main.go: description said "Read-only" but message post exists
- thread.go: dead code in error handler (both branches identical)
- reaction.go: hardcoded "message_not_found" instead of actual error

Refactoring:
- Extract generic toMap() in root.go, replace 9 identical *ToMap functions
- Move requireSessionToken from saved.go to root.go (used by section.go too)
- Remove unused encoding/json imports from 7 files
