+++ cmd/workspace.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/slack-go/slack"
 	"github.com/tammersaleh/slack-cli/internal/output"
@@ -34,9 +33,4 @@ func (c *WorkspaceInfoCmd) Run(cli *CLI) error {
 	return p.PrintMeta(output.Meta{})
 }
 
-func teamToMap(t slack.TeamInfo) map[string]any {
-	data, _ := json.Marshal(t)
-	var m map[string]any
-	_ = json.Unmarshal(data, &m)
-	return m
-}
+func teamToMap(t slack.TeamInfo) map[string]any { return toMap(t) }
